move the reader and writer to a separate package
