+++ ztype/BUILD.bazel
@@ -59,6 +59,7 @@ go_test(
     deps = [
         ":ztype",
         "//:go-zserio",
+        "//zstream",
         "@com_github_google_go_cmp//cmp",
         "@com_github_google_go_cmp//cmp/cmpopts",
     ],
